requirements.txt
# Core libraries
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8

# LlamaIndex with specific modules
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0

# Backend server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Additional tools
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
pydantic>=2.0.0

# Optional: For better text processing
tiktoken>=0.5.0